Cast wide integer to int conversion (silence msvc6 warning)

+++ generic/tclStringObj.c
@@ -33,7 +33,7 @@
  * See the file "license.terms" for information on usage and redistribution of
  * this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tclStringObj.c,v 1.125 2009/06/15 18:51:35 dgp Exp $ */
+ * RCS: @(#) $Id: tclStringObj.c,v 1.126 2009/07/01 15:06:06 patthoyts Exp $ */
 
 #include "tclInt.h"
 #include "tommath.h"
@@ -2133,9 +2133,9 @@ Tcl_AppendFormatToObj(
 		    numDigits = 1;
 		}
 		pure = Tcl_NewObj();
-		Tcl_SetObjLength(pure, numDigits);
+		Tcl_SetObjLength(pure, (int)numDigits);
 		bytes = TclGetString(pure);
-		toAppend = length = numDigits;
+		toAppend = length = (int)numDigits;
 		while (numDigits--) {
 		    int digitOffset;
